Saved Locations
includes a test button that will write some test strings to a file, and
a listview that will populate itself from the file.
in actuality, the file will be populated from saving locations in the
map

gonna change how the locations are stored in the file (from a String to
a SavedLocations structure). also need the Delete button on each row to
remove it from the file as well as the list

hello